cleanup network/layer mask logic for train/test
mask=None is the default now for the Network.from_* constructors.
That will automatically set it for the training case.
It will use the layer specific settings or the Layer constructor default.

If you set mask, that will overwrite it everywhere in all layers to use that mask.
You can do that for testing by setting mask="unity".