generic native (CPU/GPU CUDA) Theano Op interface, wip
I want to be able to easily create variants of our LSTM kernel.
I want that the Op can run on both CPU/GPU.
This is very much work-in-progress.
I copied over the existing LSTM CUDA code and started to abstract it a bit
to be able to run on both CPU and GPU (CUDA).