HDFDataset, do not make copy of full data in memory
Not sure why this was there. This has read the whole file into
memory and returned a Numpy array copy of it.
This is what we esp want to avoid with the HDFDataset.
Also, this functionality is already covered by the HDFDataset caching.
Also, this does not work if the data does not fit into memory.