NativeOp, fix FastBaumWelchOp for CPU on Linux
It seems this is not a problem on MacOSX.
It is also not a problem on Linux, if the CUDA headers
are included, and then the CPU code is used.
So in that case, the CUDA headers provide
some correct versions of these math functions.
But without them, on Linux, sth goes wrong.
To be honest, I don't fully understand the problem itself.