TaskSystem Pickler, getModNameForModDict, better
Should be much more efficient.
And also, when RETURNN is loaded as a package,
this will not have the nasty side effect of triggering all the
lazy module loaders (see __init__.py).

Not sure though if this new code is always correct,
or works in all the cases which we need.

Also see: https://stackoverflow.com/questions/56171796/

Anyway, this is only used by the Theano backend.