tf.util.dropout, share mask across recurrent loop if possible
When a layer gets moved out of a rec loop (due to our optimizations),
the behavior stays consistent to the case
when it is kept inside the loop.

Note that this potentially changes the behavior of old setups.
However, in most of our attention setups,
dropout was only applied to the output_prob layer,
which almost always got optimized out of the loop,
so it already had exactly that behavior.