SplitDimsLayer: Proper splits of axes with dynamic length
If an axis has a dynamic length, and this is a non-trivial split
(i.e. all other constant dims are not 1), we pad it to the multiples of
the product of all constant dims, via pad_to_multiples=True.

This commit also fixes that the axes for size_placeholder were incorrectly
calculated when the split axis < batch dim axis