batch norm, cleanup unused args
At least I assume they are unused.
Likely a user would not provide a tf.Tensor in a config.
Or the specified type was wrong
and this was supposed to be a float.
Anyway, I still don't see
how this would have been used potentially.
Maybe this was a relict from the Theano code conversion.