Generic LayerBase in_dim and out_dim options
#597

in_dim is checked for many common cases
for verification in LayerBase
(just that the dim exists in the input,
if there is a single input).

in_dim is also used in _ConcatInputLayer
to define the concatenated feature dim.

out_dim is further used for verification
in LayerBase (just that the dim exists in the output).

out_dim is used for layers which rely on
_base_get_out_data_from_opts,
e.g. LinearLayer
to define the output feature dim.

out_dim is used by CopyLayer
to define a potential new concatenated output feature dim.