Enforce axis description order independent (#795)
This is required to ensure that layers can reorder axes in whatever way to allow for potential optimizations. Nothing should depend on the order of axes. See the [RETURNN principles](https://github.com/rwth-i6/returnn/wiki/RETURNN-principles).

This is also for #792 to allow for an easier transition.

This introduces a new behavior version (#508).

While it probably requires changes for many configs, the changes should still be quite simple.