TF engine logdir, not cwd but None as fallback
This fallback was mostly used when RETURNN was started
for the test cases or other simple scripts.
I think the TF event logs were never really wanted in that case,
and it was annoying to get all the logs in the current dir.

In the normal case when there is some explicit model,
it would anyway take the dir of the model.

Or in all cases, it is possible to explicitly define it
via the tf_log_dir option.