keep_over_epoch, fix initial dyn shape
Remove copy_unknown_shape because this does not work
in a pure C++ environment.
Instead, use the same mechanism as we use in compile_tf_graph.py.
This uses placeholder_with_default as a trick
to create a tensor with dynamic shape but no other dependencies.
Introduces zeros_dyn_shape.