CombineLayer, ignore sparse flag
Note that before this change, sth like a + b was possible
for sparse tensors a and b, and it was adding their values,
i.e. not interpreting them as dense tensors.

With this change, now a + b is possible when a is sparse
and b is not.

E.g. when a is the index in a dim (via argmax or so)
and b is just an offset,
b would usually not be declared as sparse.