GitHub CI, fix nosetests/pytest
Error could happen when GitHub updates a minor Python version.
Then the hardcoded python binary path
in the nosetests/pytest binaries (from the cache)
would be invalid.

For nosetests, we already hacked this in an ugly way.

This cleans it up now and should solve it for both
in a much more safe and simple way.