trigger CI
The PyPI version uses the time of the last commit.
If that time was earlier than another earlier commit,
then the resulting version is lower than that from the earlier commit,
and thus the latest PyPI version might not relate
to the latest commit.
This is a quite rare situation but it can happen
when PRs are rebased and then merged.

This commit here is just to get the latest version online,
but also to raise some awareness.