ChoiceLayer in loop without dep (#1213)
This is the case for beam search on a CTC model.
The choice of a label in one frame is independent from the earlier labels.

Determining the best sequence actually is trivial in such example,
and does not require beam search at all. Just pick the argmax in each frame.

But what about beam_size then?
We expect that we get out a beam of this size.
So we need to get the N best sequences, not just the best.

How can we efficiently do that?
I think this again needs some global search.

I was thinking more about it and I think
we really want to have the same beam search algorithm as before
to get the N best.
Also, this is not an approximation of the search
but actually the exact solution.