RF TF-net-dict parameter_assign support
We also want to support conditional parameter assignments,
e.g. to support a custom straightforward batch norm implementation
in pure RF.

This makes it unfortunately quite complex.

After quite a bit of back and forth, this is what I came up with.

This also extends the rfl.Cond logic
to support extra ops.
This is anyway nice to have.

This also extends the rfl.Cond logic
to allow None results,
which makes sense when there are potential ops
with side effects like parameter_assign.

This also extends rfl.Cond by having some callbacks,
which are used for the parameter_assign logic.
This is maybe not so nice.

This also extends the rfl.Layer by having a list of usages.
This is maybe ok.

All of this is used then by the parameter_assign logic.
The logic is not really nice and easy to understand,
but I currently don't really have an easier solution.