TF WindowLayer, use opt dim order only with new behavior version
We had a case where this change of dim order
caused some setup to break.
This was because the setup was also using an old behavior version,
where the dim order had more influence.
It would not have happened with a more recent behavior version.
However, we anyway don't want to break even such old setups.