RF label_smoothed_log_prob_gradient, fix exclude_labels
The formula was wrong. It resulted in a non-normalized distribution.
Now it's fixed, and it's properly normalized,
and you get (1-smoothing) for the case of prob=1.

This is a change in behavior!
I did not introduce a new behavior_version though,
because as I see, there is so far no usage of this option
(looking at i6_experiments).