RF conv/pool/etc, stft, window, use_mask, new behavior version 23 (#1692)
`use_mask` option implemented for:

* `Conv1d.__init__` and co, `conv`
* `TransposedConv1d.__init__` an co, `transposed_conv`
* `pool`, `max_pool`, `pool1d` and co
* `stft`
* `window`
* `scatter`

`rf_use_mask: bool` config option to explicitly specify the default for `use_mask`.

If `rf_use_mask` is not specified, for older behavior versions (<=22),
we use `use_mask=False` by default, with behavior version >=23,
we use `use_mask=True` by default.
Thus this introduces the new behavior version 23.

`BatchNorm` `use_mask` is not required to be set now.
If it is not explicitly set, it is True by default.
(Previously, it was required to set this when there were dynamic dims,
so this change does not change the behavior of any existing setup.)

Fix #1691