Dim math simplify (#1755)
This rewrites and simplifies the dim math reduction/simplification logic (e.g. --a == a, a + a = 2 * a, etc).

The code should now be more straightforward and much faster.

Also, some weird edge cases are fixed, e.g. -10+(pad_left+time) was wrong when pad_left was smaller than 10, because the earlier logic reordered this to (-10+pad_left)+time, and -10+pad_left was internally doing a relu on the sizes (because of the assumption that all sizes are always positive).

It supports slightly less cases than before. Related tests (all in test_TFUtil) where cleaned up. This doesn't seem to break any of the higher-level tests (test_TFNetworkLayer etc), so I hope this does not break any existing setups. If your setup breaks (likely an old TF-based setup), please open an issue. It shouldn't be too difficult to bring the remaining missing feature back within the new implementation.