CI fix better handling for no space left (#1773)
Fix #1770.

We use some free space script to remove some dirs, some unneeded packages.

We make sure that the `~/.cache/pip` dir is removed.

Currently only for the RF-based tests, which involve both PyTorch + TensorFlow.

This PR seems to fix the current issues. But note that it seems we are still quite close to the limits. We might need to extend this script also to other tests. Also, we might need to do even more aggressive cleanup. The `~/.local` dir (containing mostly the Python packages) is now 11GB.